Add: Teste de modulo criacional em JS
