+++ public/index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="icon" href="../assets/icon-test.png" />
-    <link rel='stylesheet' href='./style.css'>
+    <link rel='stylesheet' href='../src/css/style.css'>
     <title>Home - DConvertido</title>
   </head>
 
@@ -33,7 +33,7 @@
 
       </div>
     </div>
-    <script type="module" src="./index.js"></script>
+    <script type="module" src="../src/js/index.js"></script>
   </body>
 
 </html>
