+++ src/js/pd-criacional/interfaces/valores.js
@@ -1,0 +1,3 @@
+export const Valor = {
+  valorDeInsercao: Number,
+};
