--- src/module.js
@@ -1,5 +1,0 @@
-export function funExibir(text) {
-  return (
-    console.log(`Funcionando: ${text}`)
-  );
-}
